Increased range of SIP Expires and Min-Expires headers to 0..(2**32)-1

+++ src/gov/nist/javax/sip/stack/SIPClientTransactionImpl.java
@@ -1044,7 +1044,7 @@ public class SIPClientTransactionImpl extends SIPTransactionImpl implements SIPC
       }
       this.isMapped = true;
       // Time extracted from the Expires header.
-      int expiresTime = -1;
+      long expiresTime = -1;
 
       if (sipRequest.getHeader(ExpiresHeader.NAME) != null) {
         Expires expires = (Expires) sipRequest.getHeader(ExpiresHeader.NAME);
